+++ utils/functions/data/AutoBlacklist.js
@@ -70,10 +70,9 @@ module.exports = class AutoBlacklist {
 
     check(message, bot) {
         return new Promise(async (resolve) => {
-            //if (!message.webhookId) return resolve(false);
-
             await this.get(message.guild.id).then(async (settings) => {
                 if (!settings) return resolve(false);
+
                 const channel = settings.channel;
                 if (!channel || channel !== message.channel.id) return resolve(false);